requirements.txt
Flask==2.2.2
python-dotenv==0.21.1
stripe==6.6.0
Werkzeug==2.2.2
Flask-Cors==3.0.8
